+++ gradle/projects.main.properties
@@ -30,7 +30,7 @@ polaris-quarkus-service=quarkus/service
 polaris-quarkus-server=quarkus/server
 polaris-quarkus-spark-tests=quarkus/spark-tests
 polaris-quarkus-admin=quarkus/admin
-polaris-quarkus-test-commons=quarkus/test-commons
+polaris-quarkus-commons=quarkus/commons
 polaris-quarkus-run-script=quarkus/run-script
 polaris-eclipselink=extension/persistence/eclipselink
 polaris-jpa-model=extension/persistence/jpa-model
