+++ quarkus/admin/build.gradle.kts
@@ -53,7 +53,7 @@ dependencies {
 
   implementation("org.jboss.slf4j:slf4j-jboss-logmanager")
 
-  testImplementation(project(":polaris-quarkus-test-commons"))
+  implementation(project(":polaris-quarkus-commons"))
   testFixturesApi(project(":polaris-core"))
 
   testFixturesApi(enforcedPlatform(libs.quarkus.bom))
